Adding a new bind error, moving errors to their own header file and adding comments to Registry

+++ entis/include/entis/registry.h
@@ -14,12 +14,26 @@
 #include "sparse_set.h"
 
 namespace entis
-{
+{   
+    /**
+     * Manages all the entities with their components by
+     * providing a centralized interface for:
+     * + Creating new entities.
+     * + Checking if an entity is alive.
+     * + Killing an entity (mark it as dead and delete all of its components).
+     * + Adding a new component to an entity.
+     * + Getting a specific component.
+     * + Deleting a component of an entity.
+     * + Checking if an entity has a component.
+     * + Getting all entities that have some component.
+     */ 
     class Registry
     {
 
     public:
-
+        /**
+         * Create a new empty Registry.
+         */
         Registry()
         : current_{MAX_ID},
           entities_{},
@@ -28,22 +42,57 @@ namespace entis
 
         }
 
+        /**
+         * Create a new entity.
+         * 
+         * It returns an entity that is either a recycled one
+         * (an entity that was killed) or brand new one (an entity
+         * that hasn't been used before).
+         * 
+         * @returns a new entity that has no components attached to it.
+         */ 
         inline id_t create_entity()
         {
             return (current_ == MAX_ID) ? make_new_entity() : recycle_entity();
         }
 
+        /**
+         * Check if an entity is alive (its currently available to be used).
+         * 
+         * @param entity the entity we want to test.
+         * 
+         * @return whether or not it is alive. 
+         */
         inline bool is_alive(const id_t entity) const
         {
             return entity < entities_.size() && entities_[entity] == entity;
         }
 
+        /**
+         * Mark as dead the specified alive entity and remove all 
+         * of its components.
+         * 
+         * @param entity the entity to kill.
+         */ 
         void kill_entity(const id_t entity)
-        {
-            mark_as_death(entity);
-            delete_all_components(entity);
+        {   
+            if(is_alive(entity))
+            {
+                mark_as_death(entity);
+                delete_all_components(entity);
+            }
         }
 
+        /**
+         * Check if an entity has a component T. 
+         * 
+         * 
+         * @tparam T the type of the component we want to check.
+         * 
+         * @param entity the entity we want to test.
+         * 
+         * @returns true if the entity has a component T, false otherwise.
+         */
         template <typename T>
         bool has_component(const id_t entity)
         {
@@ -52,6 +101,16 @@ namespace entis
             return manager ? manager->has_data(entity) : false;
         }
 
+        /**
+         * Get the T component of an entity if any.
+         * 
+         * @tparam T the type of the component we want to retrieve.
+         * 
+         * @param entity the entity whose component T we want to retrieve.
+         * 
+         * @returns an optional to a reference to the component T when it exists 
+         * and an empty optional otherwise.
+         */
         template <typename T>
         Component<T> get_component(const id_t entity)
         {
@@ -67,9 +126,33 @@ namespace entis
             return component;
         }
 
+        /**
+         * Associate an alive entity with a new instace of T if it 
+         * doesn't have a component associated with it already, update 
+         * the component otherwise.  
+         * 
+         * Instead of changing ownership of a component T (std::move)
+         * or making a copy of it, we pass the necessary parameters 
+         * for creating a new instance.
+         * 
+         * @tparam T the type of the component we want to bind.
+         * @tparam ...Args a packed list of values that will be 
+         * perfectly-forwarded to the constructor of T.
+         * 
+         * @param entity the entity we want to associate to the T component.
+         * 
+         * @returns an empty optional when the binding operation is successful but,
+         * it contains a BindError::INVALID_KEY when trying to bind the null entity
+         * or a BindError::DEAD_ENTITY when trying to bind a component to a dead entity.
+         */ 
         template <typename T, typename... Args>
         BindResult bind(const id_t entity, Args&&... args)
         {
+            if (!is_alive(entity))
+            {
+                return std::optional<error::BindError>{error::BindError::DEAD_ENTITY};
+            }
+
             ComponentManager<T> manager = get_component_manager<T>();
 
             if(!manager)
@@ -80,6 +163,17 @@ namespace entis
             return manager->bind(entity, args...);
         }
 
+        /**
+         * Delete the association between an entity and its component
+         * T if any.
+         * 
+         * @tparam T the type of the component whose association we want to delete.
+         *
+         * @param entity the entity whose association we want to delete.
+         * 
+         * @returns an optional containing the component T whose association
+         * we just deleted (if any).
+         */
         template <typename T>
         std::optional<T> unbind(const id_t entity)
         {
@@ -95,6 +189,14 @@ namespace entis
             return component;
         }
 
+        /**
+         * Get a list of entities that have a component T.
+         * 
+         * @tparam the type of the component we want our entities 
+         * to have.
+         * 
+         * @return a list of entities that have a component T.
+         */
         template<typename T>
         std::vector<id_t> entities_with_component()
         {
@@ -120,6 +222,12 @@ namespace entis
         std::vector<id_t> entities_;
         std::unordered_map<std::string, std::shared_ptr<IComponentManager>> component_managers_;
 
+        /**
+         * Create a brand new entity and add it to the
+         * vector of entities.
+         * 
+         * @returns the newly created entity.
+         */
         inline id_t make_new_entity()
         {
             id_t new_entity = entities_.size();
@@ -128,6 +236,12 @@ namespace entis
             return new_entity;
         }
 
+        /**
+         * Get the next entity of the implicit
+         * list of dead entities for reuse.
+         * 
+         * @returns the latest killed entity. 
+         */
         inline id_t recycle_entity()
         {   
             id_t recycled_entity = current_;
@@ -137,12 +251,23 @@ namespace entis
             return recycled_entity;
         }
 
+        /**
+         * Add the specified entity to the implicit list
+         * of dead entities.
+         * 
+         * @param entity the entity we want to kill.
+         */
         inline void mark_as_death(const id_t entity)
         {
             entities_[entity] = current_;
             current_ = entity;
         }
 
+        /**
+         * Deletes all the components of the specified entity.
+         * 
+         * @param entity the entity whose component to delete.
+         */
         void delete_all_components(const id_t entity)
         {
             for(auto& entry : component_managers_)
@@ -151,6 +276,14 @@ namespace entis
             }
         }
 
+        /**
+         * Get the corresponding manager for a component T.
+         * 
+         * @tparam T the type whose manager we want to retrieve
+         * 
+         * @returns a shared_ptr to the corresponding manager (SparseSet<T>),
+         * if no manager has been created the pointer is empty.
+         */
         template <typename T>
         inline ComponentManager<T> get_component_manager()
         {
@@ -166,6 +299,14 @@ namespace entis
             return sparse_set;
         }
 
+        /**
+         * Make a new component manager (SparseSet<T>).
+         * 
+         * @tparam T the type of the component that the manager 
+         * will administrate.
+         * 
+         * @returns the newly created manager.
+         */
         template <typename T>
         inline ComponentManager<T> make_component_manager()
         {
